chore: moved e2e user and session creation to setup function
