+++ drizzle/seed.ts
@@ -7,14 +7,6 @@ import "dotenv/config";
 
 import { drizzle, type PostgresJsDatabase } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
-import {
-  E2E_USER_ONE_EMAIL,
-  E2E_USER_ONE_ID,
-  E2E_USER_TWO_EMAIL,
-  E2E_USER_TWO_ID,
-  E2E_USER_ONE_SESSION_ID,
-  E2E_USER_TWO_SESSION_ID,
-} from "../e2e/constants";
 
 const DATABASE_URL = process.env.DATABASE_URL || "";
 
@@ -118,62 +110,6 @@ ${chance.paragraph()}
     return users;
   };
 
-  const seedE2EUser = async (email: string, id: string, name: string) => {
-    const [existingE2EUser] = await db
-      .selectDistinct()
-      .from(user)
-      .where(eq(user.id, id));
-
-    if (existingE2EUser) {
-      console.log("E2E Test user already exists. Skipping creation");
-      return existingE2EUser;
-    }
-
-    const userData = {
-      id: id,
-      username: `${name.split(" ").join("-").toLowerCase()}-${chance.integer({
-        min: 0,
-        max: 999,
-      })}`,
-      name,
-      email,
-      image: `https://robohash.org/${encodeURIComponent(name)}?bgset=bg1`,
-      location: chance.country({ full: true }),
-      bio: chance.sentence({ words: 10 }),
-      websiteUrl: chance.url(),
-    };
-    const [createdUser] = await db.insert(user).values(userData).returning();
-    return createdUser;
-  };
-
-  const seedE2EUserSession = async (userId: string, sessionToken: string) => {
-    const [existingE2EUserSession] = await db
-      .selectDistinct()
-      .from(session)
-      .where(eq(session.sessionToken, sessionToken));
-
-    if (existingE2EUserSession) {
-      console.log("E2E Test session already exists. Skipping creation");
-      return existingE2EUserSession;
-    }
-
-    try {
-      const currentDate = new Date();
-
-      return await db
-        .insert(session)
-        .values({
-          userId,
-          sessionToken,
-          // Set session to expire in 6 months.
-          expires: new Date(currentDate.setMonth(currentDate.getMonth() + 6)),
-        })
-        .returning();
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
   const userData = generateUserData();
 
   const addUserData = async () => {
@@ -249,19 +185,6 @@ ${chance.paragraph()}
 
     try {
       await addUserData();
-      const userOne = await seedE2EUser(
-        E2E_USER_ONE_EMAIL,
-        E2E_USER_ONE_ID,
-        "E2E Test User One",
-      );
-      const userTwo = await seedE2EUser(
-        E2E_USER_TWO_EMAIL,
-        E2E_USER_TWO_ID,
-        "E2E Test User Two",
-      );
-
-      await seedE2EUserSession(userOne.id, E2E_USER_ONE_SESSION_ID);
-      await seedE2EUserSession(userTwo.id, E2E_USER_TWO_SESSION_ID);
     } catch (error) {
       console.log("Error:", error);
     }
